fix(types): move orchestration contracts to canonical package
